Made use of eslint 6
